+++ CHANGELOG.md
@@ -3,3 +3,6 @@ Initial release
 
 ## v0.2.0
 Update to use TypeScript
+
+## v0.2.1
+Added documentation https://personaspace.github.io/js-context.
