+++ docs/assets/js/search.js
@@ -1,0 +1,3 @@
+var typedoc = typedoc || {};
+            typedoc.search = typedoc.search || {};
+            typedoc.search.data = {"kinds":{"1":"External module","64":"Function","256":"Interface","1024":"Property"},"rows":[{"id":0,"kind":1,"name":"\"index\"","url":"modules/_index_.html","classes":"tsd-kind-external-module"},{"id":1,"kind":256,"name":"ContextInterface","url":"interfaces/_index_.contextinterface.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"index\""},{"id":2,"kind":256,"name":"DataInterface","url":"interfaces/_index_.datainterface.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"index\""},{"id":3,"kind":256,"name":"DataContextInterface","url":"interfaces/_index_.datacontextinterface.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"index\""},{"id":4,"kind":1024,"name":"data","url":"interfaces/_index_.datacontextinterface.html#data","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"index\".DataContextInterface"},{"id":5,"kind":1024,"name":"context","url":"interfaces/_index_.datacontextinterface.html#context","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"index\".DataContextInterface"},{"id":6,"kind":64,"name":"makeDataContext","url":"modules/_index_.html#makedatacontext","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"index\""},{"id":7,"kind":64,"name":"makeContextObject","url":"modules/_index_.html#makecontextobject","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"index\""},{"id":8,"kind":64,"name":"makeContextKey","url":"modules/_index_.html#makecontextkey","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"index\""},{"id":9,"kind":64,"name":"resolveRootReference","url":"modules/_index_.html#resolverootreference","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"index\""},{"id":10,"kind":64,"name":"resolveDataContext","url":"modules/_index_.html#resolvedatacontext","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"index\""}]};
